Added results counter
- Changed type of totalItems in GitRepoDtoListPaginated from int to long;
- Added custom JPA query for counting number of results for provided parameters.
  The reason as to why a separate query is required as opposed to just counting number of
  items in the result list is due to performance.
- Said query is part of the GitRepoRepositoryCustom interface, and is used by the
  GitRepoService
- Added separate row to the result table to display the total number of results

+++ src/main/java/com/dabico/gseapp/dto/GitRepoDtoListPaginated.java
@@ -13,7 +13,7 @@ public class GitRepoDtoListPaginated {
     String next;
     String csvLink;
     String jsonLink;
-    int totalItems;
+    long totalItems;
     @Builder.Default
     List<GitRepoDto> items = new ArrayList<>();
 }
